Using float for video duration (#1146)

+++ funnel/models/video_mixin.py
@@ -124,13 +124,13 @@ class VideoMixin:
         raise VideoException("No video source or ID to create a cache key")
 
     @property
-    def _video_cache(self) -> Dict[str, Union[str, int, datetime]]:
+    def _video_cache(self) -> Dict[str, Union[str, float, datetime]]:
         data = redis_store.hgetall(self.video_cache_key)
         if data:
             if 'uploaded_at' in data and data['uploaded_at']:
                 data['uploaded_at'] = parse_isoformat(data['uploaded_at'], naive=False)
             if 'duration' in data and data['duration']:
-                data['duration'] = int(data['duration'])
+                data['duration'] = float(data['duration'])
         return data
 
     @_video_cache.setter
@@ -138,8 +138,8 @@ class VideoMixin:
         copied_data = data.copy()
         if copied_data['uploaded_at']:
             copied_data['uploaded_at'] = copied_data['uploaded_at'].isoformat()
-        if copied_data['duration'] and not isinstance(copied_data['duration'], int):
-            copied_data['duration'] = int(copied_data['duration'])
+        if copied_data['duration'] and not isinstance(copied_data['duration'], float):
+            copied_data['duration'] = float(copied_data['duration'])
         redis_store.hmset(self.video_cache_key, copied_data)
 
         # if video exists at source, cache for 2 days, if not, for 6 hours
@@ -149,7 +149,7 @@ class VideoMixin:
     # TODO: Create a dataclass for data
 
     @property
-    def video(self) -> Optional[Dict[str, Union[str, int, datetime]]]:
+    def video(self) -> Optional[Dict[str, Union[str, float, datetime]]]:
         data = None
         if self.video_source and self.video_id:
             # Check for cached data
@@ -161,7 +161,7 @@ class VideoMixin:
                     'id': self.video_id,
                     'url': cast(str, self.video_url),
                     'embeddable_url': cast(str, self.embeddable_video_url),
-                    'duration': 0,
+                    'duration': 0.0,
                     'uploaded_at': '',
                     'thumbnail': '',
                 }
